# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flights API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/reservations": {
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Create a reservation",
                "description": "Stores a new reservation along with itineraries.",
                "operationId": "storeReservation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "origin",
                                    "destination",
                                    "departure_time",
                                    "itineraries"
                                ],
                                "properties": {
                                    "origin": {
                                        "description": "Origin location",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "JFK"
                                    },
                                    "destination": {
                                        "description": "Destination location",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "LAX"
                                    },
                                    "departure_time": {
                                        "description": "Departure time",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-09-18T10:30:00Z"
                                    },
                                    "itineraries": {
                                        "description": "List of itineraries",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "origin",
                                                "destination",
                                                "departure_time",
                                                "arrival_time"
                                            ],
                                            "properties": {
                                                "origin": {
                                                    "description": "Origin location for itinerary",
                                                    "type": "string",
                                                    "maxLength": 255,
                                                    "example": "JFK"
                                                },
                                                "destination": {
                                                    "description": "Destination location for itinerary",
                                                    "type": "string",
                                                    "maxLength": 255,
                                                    "example": "LAX"
                                                },
                                                "departure_time": {
                                                    "description": "Departure time for itinerary",
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-09-18T10:30:00Z"
                                                },
                                                "arrival_time": {
                                                    "description": "Arrival time for itinerary",
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-09-18T13:30:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "reserve": {
                                            "properties": {
                                                "origin": {
                                                    "type": "string",
                                                    "example": "JFK"
                                                },
                                                "destination": {
                                                    "type": "string",
                                                    "example": "LAX"
                                                },
                                                "departure_time": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-09-18T10:30:00Z"
                                                },
                                                "itineraries": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "origin": {
                                                                "type": "string",
                                                                "example": "JFK"
                                                            },
                                                            "destination": {
                                                                "type": "string",
                                                                "example": "LAX"
                                                            },
                                                            "departure_time": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2024-09-18T10:30:00Z"
                                                            },
                                                            "arrival_time": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2024-09-18T13:30:00Z"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}